Move  skeleton objects to own folder
