+++ editor/objects/skeleton/SkeletonBone.java
@@ -1,4 +1,4 @@
-package editor.objects;
+package editor.objects.skeleton;
 
 import java.util.Enumeration;
